refactor(components): RegisterExpenses removed and EditExpenses improved
