+++ src/EditExpense.css
@@ -1,66 +1,192 @@
-.editExpenseForm__value {
-    background-color: var(--color-background-light);
-    border-radius: var(--border-radius-default);
-    display: grid;
-    grid-template-columns: auto 1fr;
-    padding: 3rem 1rem;
+fieldset,
+.newExpenseForm__value,
+.newExpenseForm__description,
+.newExpenseForm__expiration {
+    background: #ffffff;
 }
 
-.editExpenseForm__value > span:first-child {
-    color: var(--color-danger);
-    margin-top: 0.5rem;
-}
-
-.editExpenseForm__value > input {
-    max-width: 100%;
-    text-align: center;
-    width: 100%;
-}
-
-.editExpenseForm__fields {
-    background-color: var(--color-background-light);
-    border-radius: var(--border-radius-default);
-    display: grid;
-    gap: var(--spacing-small);
-    grid-template-rows: repeat(4, auto);
-    grid-template-columns: 1fr 1fr;
-    margin-top: var(--spacing-default);
-    padding: var(--spacing-default);
-}
-
-.editExpenseForm__fields label > span:first-child,
-.editExpenseForm__fields legend {
-    display: block;
-    font-weight: 700;
-    margin-bottom: var(--spacing-extra-small);
-}
-
-.editExpenseForm__fields input[type="text"] {
-    background-color: var(--color-background-gray);
-    border-radius: var(--border-radius-default);
-    padding: var(--spacing-small);
-    width: 100%;
-}
-
-#dueDate {
-    background-color: var(--color-background-light);
-    padding: 0;
-    text-align: center;
-}
-
-.editExpenseForm__fieldset {
+fieldset label {
+    align-items: center;
     display: flex;
-    flex-direction: column;
-    gap: var(--spacing-extra-small);
+    justify-content: flex-start;
+    margin-left: 0.7rem;
 }
 
-.editExpenseForm__actionButton {
-    display: block;
-    padding: var(--spacing-small);
-    width: 100%;
+fieldset label:nth-child(2) {
+    margin-top: 1.8rem;
+}
+
+fieldset label:not(:last-child) {
+    margin-bottom: 1.3rem;
+}
+
+fieldset label span {
+    margin-left: 0.9rem;
+}
+
+label,
+strong {
     font-size: 1.6rem;
 }
 
-.editExpenseForm__actionButton:not(:first-of-type) {
-    margin-top: var(--spacing-extra-small);
+strong {
+    display: block;
+    font-weight: 700;
+}
+
+input[type="radio"],
+input[type="checkbox"] {
+    opacity: 0;
+}
+
+input::-webkit-calendar-picker-indicator {
+    display: none;
+}
+
+input::-webkit-datetime-edit-fields-wrapper {
+    display: flex;
+    justify-content: center;
+}
+
+.newExpenseContainer {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    grid-template-rows: repeat(auto, 2);
+}
+
+.newExpenseHeader {
+    align-items: center;
+    display: flex;
+    justify-content: space-between;
+}
+
+.newExpenseHeader h1 {
+    color: #444c66;
+    font-size: 3.6rem;
+    font-weight: 400;
+}
+
+.newExpenseForm {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    grid-template-rows: repeat(7, auto);
+}
+
+.newExpenseForm__value,
+.newExpenseForm__value input,
+.newExpenseForm__expirationdate {
+    color: #ff7e78;
+}
+
+.newExpenseForm__value {
+    border-radius: 0.6rem;
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    height: 8rem;
+    justify-items: end;
+    margin-top: 1.3rem;
+}
+
+.newExpenseForm__value sup {
+    font-size: 1.4rem;
+    margin-top: 2.4rem;
+}
+
+.newExpenseForm__value input {
+    background: none;
+    font-size: 3.6rem;
+    width: 100%;
+}
+
+.newExpenseForm__description {
+    border-top-left-radius: 0.6rem;
+    border-top-right-radius: 0.6rem;
+    margin-top: 1.5rem;
+    padding: 2.2rem 4.5rem 1.8rem 2.2rem;
+}
+
+.newExpenseForm__description input {
+    background: #ebebeb;
+    height: 4.3rem;
+    margin-top: 1.1rem;
+    width: 100%;
+    padding-left: 0.9rem;
+}
+
+.newExpenseForm__categories {
+    padding: 0rem 4.1rem 2.3rem 2.2rem;
+}
+
+.newExpenseForm__categories label,
+.newExpenseForm__type label,
+.newExpenseForm__expiration {
+    position: relative;
+}
+
+.newExpenseForm__categories label span::after,
+.newExpenseForm__type label span::after {
+    border-radius: 50%;
+    left: 0;
+}
+
+.newExpenseForm__expiration span::after {
+    left: 2.2rem;
+}
+
+.newExpenseForm__categories label span::after,
+.newExpenseForm__type label span::after,
+.newExpenseForm__expiration span::after {
+    background-color: #f1eeee;
+    border: 1.5px solid #cccccc;
+    content: "";
+    height: 12px;
+    pointer-events: none;
+    position: absolute;
+    top: 0.3rem;
+    width: 12px;
+}
+
+.newExpenseForm__categories label input:checked + span::after,
+.newExpenseForm__type label input:checked + span::after,
+.newExpenseForm__expiration input:checked + span::after {
+    background-color: #01c9b8;
+}
+
+.newExpenseForm__expiration {
+    padding-left: 2.2rem;
+    padding-bottom: 1.5rem;
+}
+
+.newExpenseForm__expiration span {
+    margin-left: 0.7rem;
+}
+
+.newExpenseForm__expirationdate {
+    border-bottom-left-radius: 0.6rem;
+    border-bottom-right-radius: 0.6rem;
+    font-size: 3.6rem;
+    padding-bottom: 1rem;
+}
+
+.newExpenseForm__button {
+    color: #ffffff;
+    font-size: 1.6rem;
+    height: 3.5rem;
+    margin-bottom: var(--spacing-extra-small);
+}
+
+.newExpenseForm__button:hover {
+    filter: brightness(0.95);
+}
+
+.newExpenseForm__button--primary {
+    background-color: #01c9b8;
+}
+
+.newExpenseForm__button--secondary {
+    background-color: #a2a1a1;
+}
+
+.twoColumns {
+    grid-column: 1/3;
 }
